requirements.txt
fastapi
numpy
pytest
pytest-cov
python-jose[cryptography]
python-dotenv